# openapi.yaml
openapi: 3.0.3
info:
  title: Kanban Simple API
  description: |
    A lightweight kanban board REST API for task management and bot integration.
    Built with Go, SQLite, and designed for simplicity and performance.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/yourusername/kanban-simple
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://kanban.yourdomain.com/api
    description: Production server
...
paths:
...
  /cards/{cardId}/comments:
    get:
      tags:
        - Comments
      summary: Get card comments
      description: Retrieve all comments for a card
      operationId: getCardComments
      parameters:
        - $ref: '#/components/parameters/cardId'
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Comments
      summary: Add comment
      description: Add a comment to a card
      operationId: addComment
      parameters:
        - $ref: '#/components/parameters/cardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /cards/{cardId}/labels:
    get:
      tags:
        - Labels
      summary: Get card labels
      description: Get all labels assigned to a card
      operationId: getCardLabels
      parameters:
        - $ref: '#/components/parameters/cardId'
      responses:
        '200':
          description: List of labels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
...
    cardId:
      name: cardId
      in: path
      required: true
      description: Card ID
      schema:
        type: integer
        minimum: 1
...
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        card_id:
          type: integer
          example: 1
        content:
          type: string
          example: "This task is blocked waiting for API design approval"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - card_id
        - content
        - created_at
    CreateCommentRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 5000
          example: "This task is blocked waiting for API design approval"
      required:
        - content
    Label:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Bug"
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: "#ef4444"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - created_at
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid request body"
      required:
        - error
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                error: "Bad Request"
                message: "title is required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                error: "Not Found"
                message: "Card not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            dbError:
              value:
                error: "Internal Server Error"
                message: "Database connection failed"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Optional JWT authentication for bot integration (not currently implemented)"
security: []  # No authentication required by default